requirements.txt
# ICR-Net Requirements
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=8.3.0
h5py>=3.1.0
tqdm>=4.62.0
pyyaml>=5.4.0
tensorboard>=2.7.0

# Data processing
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

